+++ pkg/util/k8sutil/pod_util.go
@@ -35,11 +35,11 @@ func etcdVolumeMounts() []v1.VolumeMount {
 	}
 }
 
-func etcdContainer(cmd []string, baseImage, version string) v1.Container {
+func etcdContainer(cmd []string, repo, version string) v1.Container {
 	c := v1.Container{
 		Command: cmd,
 		Name:    "etcd",
-		Image:   ImageName(baseImage, version),
+		Image:   ImageName(repo, version),
 		Ports: []v1.ContainerPort{
 			{
 				Name:          "server",